Added functionality to gate widget- it loads a json file now
