+++ GUI/Src/exampleGateMission.json
@@ -1,0 +1,14 @@
+{
+	"type":"Gate_No_Vision",
+	"name":"gate_no_vision",
+	"timeout":2,
+	"line_up_position":[148, -20, 9, 5],
+	"position_buffer_zone":2.0,
+	"depth_buffer_zone":0.1,
+	"yaw_buffer_zone":5,
+	"stabilization_time":10.5,
+	"move_forward_dist":15,
+	"go_through_gate_backwards":true
+
+    }
+
